将MediaSinkDelegate和Demux移到MediaSink中
