+++ src/Common/MediaSink.cpp
@@ -10,6 +10,7 @@
 
 #include "MediaSink.h"
 #include "Extension/AAC.h"
+#include "Common/config.h"
 
 using namespace std;
 
@@ -271,4 +272,86 @@ void MediaSink::enableMuteAudio(bool flag) {
     _add_mute_audio = flag;
 }
 
+///////////////////////////DemuxerSink//////////////////////////////
+
+void MediaSinkDelegate::setTrackListener(TrackListener *listener) {
+    _listener = listener;
+}
+
+bool MediaSinkDelegate::onTrackReady(const Track::Ptr &track) {
+    if (_listener) {
+        _listener->addTrack(track);
+    }
+    return true;
+}
+
+void MediaSinkDelegate::onAllTrackReady() {
+    if (_listener) {
+        _listener->addTrackCompleted();
+    }
+}
+
+void MediaSinkDelegate::resetTracks() {
+    MediaSink::resetTracks();
+    if (_listener) {
+        _listener->resetTracks();
+    }
+}
+
+///////////////////////////Demuxer//////////////////////////////
+
+void Demuxer::setTrackListener(TrackListener *listener, bool wait_track_ready) {
+    if (wait_track_ready) {
+        auto sink = std::make_shared<MediaSinkDelegate>();
+        sink->setTrackListener(listener);
+        _sink = std::move(sink);
+    }
+    _listener = listener;
+}
+
+bool Demuxer::addTrack(const Track::Ptr &track) {
+    if (!_sink) {
+        _origin_track.emplace_back(track);
+        return _listener ? _listener->addTrack(track) : false;
+    }
+
+    if (_sink->addTrack(track)) {
+        track->addDelegate([this](const Frame::Ptr &frame) {
+            return _sink->inputFrame(frame);
+        });
+        return true;
+    }
+    return false;
+}
+
+void Demuxer::addTrackCompleted() {
+    if (_sink) {
+        _sink->addTrackCompleted();
+    } else if (_listener) {
+        _listener->addTrackCompleted();
+    }
+}
+
+void Demuxer::resetTracks() {
+    if (_sink) {
+        _sink->resetTracks();
+    } else if (_listener) {
+        _listener->resetTracks();
+    }
+}
+
+vector<Track::Ptr> Demuxer::getTracks(bool ready) const {
+    if (_sink) {
+        return _sink->getTracks(ready);
+    }
+
+    vector<Track::Ptr> ret;
+    for (auto &track : _origin_track) {
+        if (ready && !track->ready()) {
+            continue;
+        }
+        ret.emplace_back(track);
+    }
+    return ret;
+}
 }//namespace mediakit
