+++ src/Record/MP4Recorder.h
@@ -13,17 +13,14 @@
 
 #include <mutex>
 #include <memory>
-#include "Player/PlayerBase.h"
-#include "Util/util.h"
-#include "Util/logger.h"
-#include "Util/TimeTicker.h"
-#include "Util/TimeTicker.h"
 #include "Common/MediaSink.h"
-#include "MP4Muxer.h"
+#include "Record/Recorder.h"
 
 namespace mediakit {
 
 #ifdef ENABLE_MP4
+class MP4Muxer;
+
 class MP4Recorder final : public MediaSinkInterface {
 public:
     using Ptr = std::shared_ptr<MP4Recorder>;
@@ -63,7 +60,7 @@ private:
     std::string _full_path;
     std::string _full_path_tmp;
     RecordInfo _info;
-    MP4Muxer::Ptr _muxer;
+    std::shared_ptr<MP4Muxer> _muxer;
     std::list<Track::Ptr> _tracks;
     uint64_t _last_dts = 0;
 };
