+++ src/Rtmp/RtmpDemuxer.cpp
@@ -7,7 +7,7 @@
  * LICENSE file in the root of the source tree. All contributing project authors
  * may be found in the AUTHORS file in the root of the source tree.
  */
-
+#include "RtmpCodec.h"
 #include "RtmpDemuxer.h"
 #include "Extension/Factory.h"
