+++ src/Rtp/RtpSender.h
@@ -14,7 +14,7 @@
 #include "PSEncoder.h"
 #include "Extension/CommonRtp.h"
 #include "Rtcp/RtcpContext.h"
-
+#include "Common/MediaSource.h"
 namespace mediakit{
 
 //rtp发送客户端，支持发送GB28181协议
