+++ src/Rtsp/RtspDemuxer.h
@@ -12,12 +12,10 @@
 #define SRC_RTP_RTSPDEMUXER_H_
 
 #include <unordered_map>
-#include "Player/PlayerBase.h"
-#include "Util/TimeTicker.h"
-#include "RtpCodec.h"
+#include "Common/MediaSink.h"
 
 namespace mediakit {
-
+class RtpCodec;
 class RtspDemuxer : public Demuxer {
 public:
     typedef std::shared_ptr<RtspDemuxer> Ptr;
@@ -51,8 +49,8 @@ private:
     float _duration = 0;
     AudioTrack::Ptr _audio_track;
     VideoTrack::Ptr _video_track;
-    RtpCodec::Ptr _audio_rtp_decoder;
-    RtpCodec::Ptr _video_rtp_decoder;
+    std::shared_ptr<RtpCodec> _audio_rtp_decoder;
+    std::shared_ptr<RtpCodec> _video_rtp_decoder;
 };
 
 } /* namespace mediakit */
